Extract `native` package
