+++ lib/NotepadType.dart
@@ -1,20 +1,22 @@
 import 'package:tuple/tuple.dart';
 
 import 'NotepadClient.dart';
-
-enum BleInputProperty { Disabled, Indication }
+import 'native/BleType.dart';
 
 class NotepadType {
   final NotepadClient _notepadClient;
+
+  final _bleType = BleType();
 
   NotepadType(this._notepadClient);
 
   void configCharacteristics() {
     for (var serviceCharacteristic in _notepadClient.inputIndicationCharacteristics)
-      configInputCharacteristic(serviceCharacteristic, BleInputProperty.Indication);
+      configInputCharacteristic(serviceCharacteristic, BleInputProperty.indication);
   }
 
   void configInputCharacteristic(Tuple2<String, String> serviceCharacteristic, BleInputProperty inputProperty) {
     print('configInputCharacteristic $serviceCharacteristic, $inputProperty');
+    _bleType.setNotifiable(serviceCharacteristic, inputProperty);
   }
 }
